Remove extreneous `image` features and dependencies (#1425)
* Cut out all image dependencies except for png and jpeg

* Small refactor of re_log: split out a file

* deny.toml: remove recursive exceptions for `ring` and `winit`

* ci: cargo deny check all four target platforms

* fix cargo deny CI snafu

* Still I messed up the CI command
