+++ deny.toml
@@ -39,18 +39,17 @@ deny = [
 ]
 
 skip = [
-  { name = "ahash" },   # Popular crate + fast release schedule = lots of crates still using old versions
-  { name = "base64" },  # Too popular
-  { name = "nix" },     # Too many to protect against
-  { name = "time" },    # Too popular
-  { name = "windows" }, # wgpu uses newer version than rfd, which uses same on latest but isnt' released yet https://github.com/PolyMeilex/rfd
+  { name = "ahash" },     # Popular crate + fast release schedule = lots of crates still using old versions
+  { name = "base64" },    # Too popular
+  { name = "memoffset" }, # Small crate
+  { name = "nix" },       # Too many to protect against
+  { name = "time" },      # Too popular
+  { name = "windows" },   # wgpu and rfd are sometimes on different versions
 ]
 skip-tree = [
   { name = "cargo-run-wasm" }, # Dev-tool
   { name = "comfy-table" },    # arrow vs. polars use different major versions (polars' one is dev-dep only)
   { name = "criterion" },      # dev-dependency
-  { name = "ring" },           # The TLS stack imports a million things.
-  { name = "winit" },          # winit pulls in a lot of old, duplicated crates. At least it is not wasm
 ]
