+++ examples/rust/objectron/Cargo.toml
@@ -13,7 +13,7 @@ rerun = { workspace = true, features = ["web_viewer"] }
 anyhow.workspace = true
 clap = { workspace = true, features = ["derive"] }
 glam.workspace = true
-image = { version = "0.24", default-features = false, features = ["jpeg"] }
+image = { workspace = true, default-features = false, features = ["jpeg"] }
 itertools = "0.10"
 prost = "0.11"
