+++ scripts/check.sh
@@ -40,8 +40,9 @@ cargo doc --document-private-items --no-deps --all-features
 
 ./scripts/lint.py
 
-cargo deny --all-features --log-level error check
+cargo deny --all-features --log-level error --target aarch64-apple-darwin check
 cargo deny --all-features --log-level error --target wasm32-unknown-unknown check
+cargo deny --all-features --log-level error --target x86_64-pc-windows-msvc
 cargo deny --all-features --log-level error --target x86_64-unknown-linux-musl check
 
 echo "All checks passed!"
